Revert "KAFKA-15764: Missing Tests for Transactions (#14702)" (#15029)
This reverts commit b998f6a.

We have been seeing a lot of failures of TransactionsWithTieredStoreTest.testTransactionsWithCompression on trunk, and it seems to start with this PR. I see how this PR can influence the test via the change in TestUtils. The bad part is that sometimes seems to kill the Gradle Executors completely. So I'd suggest reverting the change before investigating further to stabilize CI.

Reviewers: Bruno Cadonna <cadonna@apache.org>